+++ Sources/Scheduler.swift
@@ -111,8 +111,8 @@ public final class UIScheduler: Scheduler {
 	}()
 
 	deinit {
-		queueLength.deinitialize()
-		queueLength.deallocate(capacity: 1)
+		queueLength.deinitialize(count: 1)
+		queueLength.deallocate()
 	}
 	#endif