+++ Tests/ReactiveSwiftTests/SignalSpec.swift
@@ -1012,7 +1012,7 @@ class SignalSpec: QuickSpec {
 
 			it("should skip values according to a predicate") {
 				let (baseSignal, observer) = Signal<String, NoError>.pipe()
-				let signal = baseSignal.skipRepeats { $0.characters.count == $1.characters.count }
+				let signal = baseSignal.skipRepeats { $0.count == $1.count }
 
 				var values: [String] = []
 				signal.observeValues { values.append($0) }
@@ -3594,10 +3594,10 @@ class SignalSpec: QuickSpec {
 			describe("log events") {
 				it("should output the correct event without identifier") {
 					let expectations: [(String) -> Void] = [
-						{ event in expect(event) == "[] value 1" },
-						{ event in expect(event) == "[] completed" },
-						{ event in expect(event) == "[] terminated" },
-						{ event in expect(event) == "[] disposed" },
+						{ event in expect(event).to(equal("[] value 1")) },
+						{ event in expect(event).to(equal("[] completed")) },
+						{ event in expect(event).to(equal("[] terminated")) },
+						{ event in expect(event).to(equal("[] disposed")) },
 					]
 
 					let logger = TestLogger(expectations: expectations)
@@ -3613,10 +3613,10 @@ class SignalSpec: QuickSpec {
 
 				it("should output the correct event with identifier") {
 					let expectations: [(String) -> Void] = [
-						{ event in expect(event) == "[test.rac] value 1" },
-						{ event in expect(event) == "[test.rac] failed error1" },
-						{ event in expect(event) == "[test.rac] terminated" },
-						{ event in expect(event) == "[test.rac] disposed" },
+						{ event in expect(event).to(equal("[test.rac] value 1")) },
+						{ event in expect(event).to(equal("[test.rac] failed error1")) },
+						{ event in expect(event).to(equal("[test.rac] terminated")) },
+						{ event in expect(event).to(equal("[test.rac] disposed")) },
 					]
 
 					let logger = TestLogger(expectations: expectations)